\name{davidRian-package}
\alias{davidRian-package}
\alias{davidRian}
\docType{package}
\title{
\packageTitle{davidRian}
}
\description{
\packageDescription{davidRian}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{davidRian}
\packageIndices{davidRian}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{davidRian}

Maintainer: \packageMaintainer{davidRian}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
