% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dc_grad}
\alias{dc_grad}
\title{Gradient of the log-likelihood of univariate Davidian curves}
\usage{
dc_grad(x, phi)
}
\arguments{
\item{x}{A vector of observations.}

\item{phi}{Davidian curve parameters. length(phi) < 11.}
}
\description{
Provides the gradient for use in estimation.
}
