% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcdensC}
\alias{dcdensC}
\title{Density function for Davidian curves}
\usage{
dcdensC(x, k, mean, sd, phi)
}
\arguments{
\item{x}{An integer vector}
}
\description{
Returns the density for a vector of x.
}
