% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcGrad}
\alias{dcGrad}
\title{Gradient of the log-likelihood of a univariate Davidian curve}
\usage{
dcGrad(x, phi)
}
\arguments{
\item{x}{A vector of observations.}

\item{phi}{Davidian curve parameters. length(phi) < 11.}
}
\description{
Provides the gradient for use in estimation.
}
