% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdc}
\alias{rdc}
\title{Random samples from univariate Davidian curves}
\usage{
rdc(n, phi)
}
\arguments{
\item{n}{Number of observations to be sampled.}

\item{phi}{Davidian curve parameters. length(phi) < 11.}
}
\description{
Returns n samples from a univariate Davidian curve.
}
