% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcGrad}
\alias{dcGrad}
\title{Gradient of the log likelihood of a univariate DC}
\usage{
dcGrad(x, phi)
}
\arguments{
\item{x}{A vector of observations.}

\item{phi}{DC parameters as introduced in Woods & Lin.}
}
\description{
Gradient of the loglikelihood of a univariate DC, to be used in estimation.
}
