% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddc}
\alias{ddc}
\title{Density function for univariate Davidian curves}
\usage{
ddc(x, phi)
}
\arguments{
\item{x}{vector of quantiles.}

\item{phi}{Davidian curve parameters. length(phi) < 11.}
}
\description{
Returns the density for a vector of x.
}
