% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddc}
\alias{ddc}
\title{Density function for univariate Davidian Curves (DC)}
\usage{
ddc(x, mean, sd, phi)
}
\arguments{
\item{x}{A vector of observations.}

\item{mean}{Mean of the normal base of DC, see the package vignette.}

\item{sd}{SD of the normal base of DC, see the package vignette.}

\item{phi}{DC parameters as introduced in Woods & Lin.}
}
\description{
Returns the density for a vector of x.
}
