% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdc}
\alias{rdc}
\title{Random samples from a univariate Davidian Curve (DC)}
\usage{
rdc(n, mean, sd, phi)
}
\arguments{
\item{n}{Number of observations to be sampled.}

\item{mean}{Mean of the normal base of DC, see the package vignette.}

\item{sd}{SD of the normal base of DC, see the package vignette.}

\item{phi}{DC parameters as introduced in Woods & Lin.}
}
\description{
Returns n samples from a univariate DC.
}
